wip: On post creation the user_id will be added in order to map posts with a user
 - Added user as an argument in all the endpoints and services

+++ src/app/api/v1/endpoints/posts.py
@@ -13,29 +13,33 @@ from typing import List
 from sqlalchemy.ext.asyncio import AsyncSession
 from app.db.session import get_db
 from app.schema.post import Post, PostCreate
+from app.models.user import User
+from app.auth.service import get_current_user
 from app.services import post_service
 
 post_router = APIRouter(prefix="/posts", tags=["Posts"])
 
 
 @post_router.post("/", response_model=Post)
-async def create_post(data: PostCreate, db: AsyncSession=Depends(get_db)):
-    return await post_service.create_post(db, data)
+async def create_post(data: PostCreate, db: AsyncSession=Depends(get_db), user: User=Depends(get_current_user)):
+    return await post_service.create_post(db, data, user.id)
 
 
 @post_router.get("/", response_model=List[Post])
-async def list_posts(db: AsyncSession=Depends(get_db)):
-    return await post_service.get_posts(db)
+async def list_posts(db: AsyncSession=Depends(get_db), user: User=Depends(get_current_user)):
+    return await post_service.get_posts(db, user.id)
 
 
 @post_router.get("/{post_id}", response_model=Post)
-async def get_post(post_id: UUID, db: AsyncSession=Depends(get_db)):
-    # check for post if None raise the relevant exception or return accordingly
-    return await post_service.get_post(db, post_id)
+async def get_post(post_id: UUID, db: AsyncSession=Depends(get_db), user: User=Depends(get_current_user)):
+    if user:
+        # check for post if None raise the relevant exception or return accordingly
+        return await post_service.get_post(db, post_id)
 
 
 @post_router.delete("/{post_id}", response_model=Post)
-async def remove_post(post_id: UUID, db: AsyncSession=Depends(get_db)):
-    # check if post exists and remove
-    await post_service.delete_post(db, post_id)
-    return {"status": "deleted"}
+async def remove_post(post_id: UUID, db: AsyncSession=Depends(get_db), user: User=Depends(get_current_user)):
+    if user:
+        # check if post exists and remove
+        await post_service.delete_post(db, post_id)
+        return {"status": "deleted"}
